network-security-toolkit/requirements.txt
boto3
cryptography
dnspython
scapy
pyyaml
jinja2